Porting the SQLite methods of the picking stage to prepare, hopefully they will solve the locking issue found on the NHM cluster.

+++ Mikado/preparation/annotation_parser.py
@@ -179,23 +179,25 @@ def load_into_storage(shelf_name, exon_lines, min_length, logger, strip_cds=True
 
     """Function to load the exon_lines dictionary into the temporary storage."""
 
-    conn = sqlite3.connect(shelf_name)
-    cursor = conn.cursor()
-    try:
-        cursor.execute(
-            "CREATE TABLE dump (chrom text, start integer, end integer, strand text, tid text, features blob)")
-    except sqlite3.OperationalError:
-        # Table already exists
+    if os.path.exists(shelf_name) or any(_.startswith(os.path.basename(shelf_name))
+                                         for _ in os.listdir(os.path.dirname(shelf_name))):
         logger.error("Shelf %s already exists (maybe from a previous aborted run?), dropping its contents", shelf_name)
-        cursor.close()
-        conn.close()
         os.remove(shelf_name)
-        conn = sqlite3.connect(shelf_name)
-        cursor = conn.cursor()
-        cursor.execute(
-            "CREATE TABLE dump (chrom text, start integer, end integer, strand text, tid text, features blob)")
+        for _ in (_.startswith(os.path.basename(shelf_name)) for _ in os.listdir(os.path.dirname(shelf_name))):
+            os.remove(_)
 
+    conn = sqlite3.connect(shelf_name,
+                           isolation_level="DEFERRED",
+                           timeout=60,
+                           check_same_thread=False  # Necessary for SQLite3 to function in multiprocessing
+                           )
+    cursor = conn.cursor()
+    cursor.execute("PRAGMA journal_mode=wal")
+    cursor.execute(
+        "CREATE TABLE dump (chrom text, start integer, end integer, strand text, tid text, features blob)")
+    cursor.execute("CREATE TABLE transcripts (counter integer UNIQUE PRIMARY KEY, json BLOB)")
     cursor.execute("CREATE INDEX idx ON dump (tid)")
+    conn.commit()
     logger.debug("Created tables for shelf %s", shelf_name)
 
     for tid in exon_lines:
